requirements.txt
py-cord
python-dotenv
mysql-connector-python
asyncio